requirements.txt
alabaster==1.0.0
anyio==4.11.0
asgiref==3.10.0
babel==2.17.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docutils==0.21.2
h11==0.16.0
idna==3.11
imagesize==1.4.1
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-django==4.11.1
pytest-html==4.1.1
pytest-metadata==3.1.1
requests==2.32.5
roman-numerals-py==3.1.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-autobuild==2025.8.25
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
starlette==0.50.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
